* 'CMake' of github.com:ewmailing/IupCocoa:
  CMake: New files from Iup resync
  SymbolVisibility: iup_varg.h
  Official: revision 4693
  SymbolVisibility: Needed for va_list functions.

+++ src/iup_canvas.h
@@ -27,7 +27,7 @@ char* iupCanvasGetPosYAttrib(Ihandle* ih);
 struct _IcontrolData 
 {
   int sb;    /* scrollbar configuration, valid only after map, use iupBaseGetScrollbar before map */
-  float posx, posy;
+  double posx, posy;
   int inside_resize;
 };
